Move NumericValue, NumericNDArray to expr.numeric_expr
