+++ pyomo/core/base/indexed_component.py
@@ -20,7 +20,8 @@ import textwrap
 from copy import deepcopy
 
 from pyomo.core.expr import current as EXPR
-from pyomo.core.expr.numvalue import native_types, NumericNDArray
+from pyomo.core.expr.numeric_expr import NumericNDArray
+from pyomo.core.expr.numvalue import native_types
 from pyomo.core.base.indexed_component_slice import IndexedComponent_slice
 from pyomo.core.base.initializer import Initializer
 from pyomo.core.base.component import Component, ActiveComponent
