+++ pyomo/core/tests/unit/test_numpy_expr.py
@@ -29,7 +29,8 @@ from pyomo.environ import (
     Reals,
 )
 from pyomo.core.expr.current import MonomialTermExpression
-from pyomo.core.expr.numvalue import NumericNDArray, as_numeric
+from pyomo.core.expr.numeric_expr import NumericNDArray
+from pyomo.core.expr.numvalue import as_numeric
 from pyomo.core.expr.compare import compare_expressions
 from pyomo.core.expr.relational_expr import InequalityExpression
